Make search term highlighting optional
If the list is long, and the first few type characters match a lot of items (as it often the case) then the calls to GQuery.html would have a significant performance impact.
Make the highlighting optional (default: on) so that in use cases with long lists it can be turned off for performance reasons.