Ability to style individual items in the ChosenListBox and indent them
Ability to style individual items in the Chosen list and indent them, we have used this to show hierarchy within the ChosenListBox, and you can also add a style name to items. For example you might want the user to select from a file structure so you want the directories to be at a different level then their children, but all of the items (files and directories) are valid options for the user to select.